[Creating Controller Methods] [/controllers/bootcamps.js added] Add controllers for the bootcamps endpoint in the project. It defines six different functions that correspond to the different HTTP requests that can be made to the bootcamps endpoint.
The first three functions, getBootcamps, getBootcamp, and createBootcamp, all respond with a status code of 200 and a JSON object containing a success field that is set to true and a msg field that contains a message describing what the function does. For example, the getBootcamps function returns a message saying "Show all bootcamps." Obviously this is all pretty simple for now but as we add functionality to these controllers this provides a way of keeping our routes/bootcamp.js file clean and readable, thus routes/bootcamps will only handle our routes and our controllers/bootcamps will contain the code that defines the functionality of those routes.

The next three functions, updateBootcamp, deleteBootcamp, and deleteBootcamp, all respond with a status code of 200 and a JSON object containing a success field that is set to true and a msg field that contains a message describing what the function does. For example, the updateBootcamp function returns a message saying "Update bootcamp [id]."

Overall, this code is a set of controllers for the bootcamps endpoint that respond with appropriate status codes and messages based on the HTTP requests that are made to the endpoint. This could all be done within server.js but by doing things this way both our server.js and our routes/bootcamps are each only performing tasks appropriate to their purpose... our server is serving our routes, routes/bootcamps.js is simply defining our routes and our controllers/bootcamps is defining the behaviour of those routes.

In future updates controllers/bootcamps will become more complex, as here we will be performing our crud opperations and fetching or manipulating our bootcamps data to and from our databse so doing things this way will help to keep our code readable and importantly maintainable.

[routes/bootcamps]
The changes here are simply a refactor to make use of the changes to our file structure listed above. So we import our controllers and refactor our routes to reflect those changes.  Looking at the changelog here we have a much more clean and concise means of defining our routes.