[Third party middleware for logging: Morgan]
npm install morgan

[server.js]
import/require morgan.

add a middleware function that is used to enable logging in the development environment. The morgan package is used to log requests to the console in a specific format. In this case, the dev option is used, which logs the request method, URL, status code, response time, and the request body.

This middleware is only applied in the development environment, which can be determined by checking the NODE_ENV variable. If the value of NODE_ENV is development, the middleware is applied, otherwise it is not. This allows for different logging configurations depending on the environment.