[Connecting to mongoDB] install mongoose.
[Create db.js]

Here we create a Nodejs module that exports a function called connectDB. The function uses the mongoose library to connect to a MongoDB database. The MONGO_URI environment variable is used to store the connection string.

The code starts by importing the mongoose library. Then, the connectDB function is defined. It uses the async keyword to make the function asynchronous.

The connectDB function uses the mongoose.connect method to connect to the MongoDB database. The process.env.MONGO_URI variable is used to get the connection string from the environment.

Once the connection is established, the code logs a message to the console indicating that the MongoDB connection was successful.

Next we import the module to server.js and make a call to the function in db.js to connect to the database.

Next we change the server code so that we can mannage unhandled exceptions. here:

On line 28, the app.listen() method is now placed within the variable server and is used to start the server and listen for requests on the specified port. The callback function logs a message to the console indicating that the server is running and provides the port number.

On line 32, the process.on() method is used to listen for uncaught exceptions. If an exception is thrown and not handled, the callback function will log the error message to the console and then close the server and exit the process with a status code of 1. This will cause the Node.js process to terminate and any active connections to be closed.
We can now connect to our database. And exit if we have unhandled exceptions.
Happy hacking!!