[First Data Model] Added a folder to workspace _data which contains dummy data for our application.
Next we created a new directory models and added a Bootcamp.js file to this directory. Here  we constructed our first data model for this directory like so.

added a schema definition for a MongoDB document. It is used to define the structure of the data that will be stored in the database.

The code starts by requiring the Mongoose library, which is a popular Node.js library for working with MongoDB.

Next, the code defines a new Schema using the mongoose.Schema constructor. A Schema is a blueprint for a MongoDB document, defining the structure of the data that will be stored in the document.

The BootcampSchema defines several fields for the document, including:

name: A required string field that is unique and can be up to 50 characters long.
description: A required string field that can be up to 1000 characters long.
website: A string field that must be a valid URL.
email: A string field that must be a valid email address.
address: A required string field.
location: A GeoJSON object that contains fields for the location's coordinates, formatted address, and other details.
careers: An array of strings that must be one of several allowed values.
averageRating: A number field that must be between 1 and 10, inclusive.
averageCost: A number field.
photo: A string field that defaults to a value of "no-photo.jpg".
housing: A boolean field that defaults to false.
jobAssistance: A boolean field that defaults to false.
jobGuarantee: A boolean field that defaults to false.
acceptGi: A boolean field that defaults to false.
createdAt: A Date field that defaults to the current date and time.

The code then exports the BootcampSchema as a module, making it available for use in other parts of the application.